refactor: add controller.common package
